# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Netriun Nexus API",
    "version": "1.7.0",
    "description": "Multi-tenant workspace API with live provider refresh, database snapshots, custom account access policies, local verified-email authentication, and workspace-scoped OIDC/SAML SSO. Lists return {data: [...]}; errors return {error: message}. Login tokens expire after 12 hours. Community workspaces include one owner and five additional users. The workspace owner retains local break-glass access when SSO is required."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Sign in; returns token, user, expires_in and session cookie",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Create an isolated Community workspace and send owner email verification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Signup"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Workspace created; email verification is required before sign-in"
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Email or username already in use"
          },
          "429": {
            "description": "Sign-up rate limit exceeded"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": []
      }
    },
    "/auth/resend-verification": {
      "post": {
        "summary": "Request a replacement email-verification link",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": { "email": { "type": "string", "format": "email" } },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "202": { "description": "Generic response returned whether or not the account exists" },
          "429": { "description": "Rate limit exceeded" }
        },
        "security": []
      }
    },
    "/auth/sso/providers": {
      "get": {
        "summary": "Discover enabled identity providers using a workspace SSO code",
        "parameters": [{ "name": "workspace", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Enabled provider names, protocols, and sign-in URLs" },
          "400": { "description": "Workspace code is required" },
          "429": { "description": "Discovery rate limit exceeded" }
        },
        "security": []
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Revoke current session",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/summary": {
      "get": {
        "summary": "Scoped inventory counts",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "description": "Limit counts to one accessible cloud account",
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/accounts": {
      "get": {
        "summary": "List accounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      },
      "post": {
        "summary": "Create Account",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Account"
              }
            }
          }
        }
      }
    },
    "/accounts/test": {
      "post": {
        "summary": "Validate new cloud credentials and required compute read access",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConnectionTest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Credentials and required read access are working" },
          "400": { "description": "Invalid provider or credentials" },
          "401": { "description": "Authentication required" },
          "403": { "description": "Cloud account management access required" },
          "429": { "description": "Connection-test rate limit exceeded" },
          "502": { "description": "Provider rejected the connection; response includes a stable NX error code and remediation steps" }
        }
      }
    },
    "/accounts/{id}/test": {
      "post": {
        "summary": "Validate a saved cloud connection or replacement credentials",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConnectionTest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Stored or replacement credentials and required read access are working" },
          "400": { "description": "Invalid provider or replacement credentials" },
          "401": { "description": "Authentication required" },
          "403": { "description": "Account management access required" },
          "404": { "description": "Account not found" },
          "429": { "description": "Connection-test rate limit exceeded" },
          "502": { "description": "Provider rejected the connection; response includes a stable NX error code and remediation steps" }
        }
      }
    },
    "/accounts/{id}": {
      "put": {
        "summary": "Update Account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Account"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/groups": {
      "get": {
        "summary": "List groups",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      },
      "post": {
        "summary": "Create Group",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Group"
              }
            }
          }
        }
      }
    },
    "/groups/{id}": {
      "put": {
        "summary": "Update Group",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Group"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Group",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "List users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      },
      "post": {
        "summary": "Create User",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    },
...
    "/instances": {
      "get": {
        "summary": "List instances",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "description": "Limit inventory to one accessible cloud account",
            "schema": { "type": "integer", "minimum": 1 }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
    "/accounts/{id}/eds/regions": {
      "get": {
        "summary": "List supported WUYING EDS regions with availability, per-region counts, and the total available desktop count",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "responses": { "200": { "description": "Region list with desktop counts; unavailable counts are -1" }, "400": { "description": "Invalid provider" }, "403": { "description": "Access denied" }, "502": { "description": "Alibaba EDS rejected the request" } }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nexus_session"
      }
    },
    "schemas": {
      "Login": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "ConnectionTest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "provider": {
            "type": "string",
            "enum": ["aws", "alibaba", "azure", "gcp"]
          },
          "credentials": {
            "$ref": "#/components/schemas/Account/properties/credentials"
          }
        },
        "required": ["provider"]
      },
...
    }
  }
}